mv captureStdout as public func
