+++ pkg/app/app_test.go
@@ -5,7 +5,6 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"log"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -18,6 +17,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 
 	"github.com/roboll/helmfile/pkg/remote"
+	"github.com/roboll/helmfile/pkg/testutil"
 
 	"gotest.tools/v3/assert"
 
@@ -4668,36 +4668,6 @@ See https://github.com/roboll/helmfile/issues/878 for more information.
 	}
 }
 
-func captureStdout(f func()) string {
-	reader, writer, err := os.Pipe()
-	if err != nil {
-		panic(err)
-	}
-	stdout := os.Stdout
-	defer func() {
-		os.Stdout = stdout
-		log.SetOutput(os.Stderr)
-	}()
-	os.Stdout = writer
-	log.SetOutput(writer)
-	out := make(chan string)
-	wg := new(sync.WaitGroup)
-	wg.Add(1)
-	go func() {
-		var buf bytes.Buffer
-		wg.Done()
-		_, err = io.Copy(&buf, reader)
-		if err != nil {
-			panic(err)
-		}
-		out <- buf.String()
-	}()
-	wg.Wait()
-	f()
-	writer.Close()
-	return <-out
-}
-
 func TestPrint_SingleStateFile(t *testing.T) {
 	files := map[string]string{
 		"/path/to/helmfile.yaml": `
@@ -4726,7 +4696,7 @@ releases:
 
 	expectNoCallsToHelm(app)
 
-	out := captureStdout(func() {
+	out := testutil.CaptureStdout(func() {
 		err := app.PrintState(configImpl{})
 		assert.NilError(t, err)
 	})
@@ -4773,7 +4743,7 @@ releases:
 
 	expectNoCallsToHelm(app)
 
-	out := captureStdout(func() {
+	out := testutil.CaptureStdout(func() {
 		err := app.PrintState(configImpl{})
 		assert.NilError(t, err)
 	})
@@ -4834,7 +4804,7 @@ releases:
 
 	expectNoCallsToHelm(app)
 
-	out := captureStdout(func() {
+	out := testutil.CaptureStdout(func() {
 		err := app.ListReleases(configImpl{})
 		assert.NilError(t, err)
 	})
@@ -4896,7 +4866,7 @@ releases:
 
 	expectNoCallsToHelm(app)
 
-	out := captureStdout(func() {
+	out := testutil.CaptureStdout(func() {
 		err := app.ListReleases(configImpl{
 			output: "json",
 		})
